+++ CHANGELOG.rst
@@ -2,7 +2,7 @@
 Changelog
 =========
 
-* adapting tox to pyge project
+* Implementation of a new `singlechain` function. Goal: provide a function to return GE values directly from a PDB file
 
 v0.1.0 (2022-12-15)
 ------------------------------------------------------------